Promote Int conversion to long long int
git-svn-id: https://svn.atnf.csiro.au/difx@7172 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ libraries/dirlist/trunk/src/dirlist_parameter.cpp
@@ -24,14 +24,14 @@ bool DirListParameter::setFromTokens(const std::vector<std::string> &tokens)
 	return true;
 }
 
-int DirListParameter::getInt(unsigned int index) const
+long long int DirListParameter::getInt(unsigned int index) const
 {
 	if(index >= items.size())
 	{
 		throw DirListException("getInt(): index out of range.");
 	}
 
-	return atoi(items[index].c_str());
+	return atoll(items[index].c_str());
 }
 
 double DirListParameter::getDouble(unsigned int index) const
